+++ clients/go/ahp/reducers.go
@@ -1156,6 +1156,7 @@ func ApplyActionToChangeset(state *ahptypes.ChangesetState, action ahptypes.Stat
 		*ahptypes.ChangesetFileSetAction,
 		*ahptypes.ChangesetFileRemovedAction,
 		*ahptypes.ChangesetOperationsChangedAction,
+		*ahptypes.ChangesetOperationStatusChangedAction,
 		*ahptypes.ChangesetClearedAction:
 		return ReduceOutcomeNoOp
 	}