+++ types/common/actions.ts
@@ -64,6 +64,7 @@ import type {
   ChangesetFileSetAction,
   ChangesetFileRemovedAction,
   ChangesetOperationsChangedAction,
+  ChangesetOperationStatusChangedAction,
   ChangesetClearedAction,
 } from '../channels-changeset/actions.js';
 
@@ -140,6 +141,7 @@ export const enum ActionType {
   ChangesetFileSet = 'changeset/fileSet',
   ChangesetFileRemoved = 'changeset/fileRemoved',
   ChangesetOperationsChanged = 'changeset/operationsChanged',
+  ChangesetOperationStatusChanged = 'changeset/operationStatusChanged',
   ChangesetCleared = 'changeset/cleared',
   RootTerminalsChanged = 'root/terminalsChanged',
   RootConfigChanged = 'root/configChanged',
@@ -240,6 +242,7 @@ export type StateAction =
   | ChangesetFileSetAction
   | ChangesetFileRemovedAction
   | ChangesetOperationsChangedAction
+  | ChangesetOperationStatusChangedAction
   | ChangesetClearedAction
   | TerminalDataAction
   | TerminalInputAction